[step3] refactor: CellProperty 서브 클래스 생성을 Factory를 통해서만 하도록 수정
