+++ src/main/kotlin/domain/cell/CellPropertyFactory.kt
@@ -11,3 +11,37 @@ class CellPropertyFactory {
         }
     }
 }
+
+private data class MineCellProperty(override var state: CellState = CellState.HIDE) : CellProperty() {
+
+    override fun getSymbol(): String {
+        return if (isOpen()) MINE_SYMBOL else CLOSE_SYMBOL
+    }
+
+    override fun isMine(): Boolean {
+        return true
+    }
+
+    override fun isCleanAroundMineCount(): Boolean {
+        return false
+    }
+
+    companion object {
+        private const val MINE_SYMBOL = "*"
+    }
+}
+
+private data class NormalCellProperty(private val aroundMineCount: AroundMineCount, override var state: CellState = CellState.HIDE) : CellProperty() {
+    override fun getSymbol(): String {
+        return if (isOpen()) aroundMineCount.value.toString() else CLOSE_SYMBOL
+    }
+
+    override fun isMine(): Boolean {
+        return false
+    }
+
+    override fun isCleanAroundMineCount(): Boolean {
+        return aroundMineCount.isClean
+    }
+}
+
